YUK-4404 Cap Bank temp move control orders get messed up
More integer casts removed. Things appear to stay in order now.

+++ yukon-client/common/src/com/cannontech/database/db/capcontrol/CCFeederBankList.java
@@ -139,8 +139,8 @@ public static java.util.ArrayList getCapBanksOnFeederList(Integer feederId, java
 							feederId, 
 							new Integer( rset.getInt("DeviceID") ),
 							new Float( rset.getFloat("ControlOrder")) ,
-                            new Float (rset.getInt("CloseOrder")),
-                            new Float (rset.getInt("TripOrder")))
+                            new Float (rset.getFloat("CloseOrder")),
+                            new Float (rset.getFloat("TripOrder")))
                         );
